Introduce CAL based transparency accelerations.
Authored by Michael Vrhel.

+++ base/gxblend.h
@@ -383,15 +383,15 @@ void pdf14_rgb_cs_to_grayspot_cm(gx_device * dev, const gs_gstate *pgs,
 void pdf14_cmyk_cs_to_grayspot_cm(gx_device * dev, frac c, frac m, frac y, frac k, frac out[]);
 
 void gx_build_blended_image_row(const byte *gs_restrict buf_ptr, int planestride,
-                                int width, int num_comp, byte bg, byte *gs_restrict linebuf);
+                                int width, int num_comp, uint16_t bg, byte *gs_restrict linebuf);
 void gx_build_blended_image_row16(const byte *gs_restrict buf_ptr, int planestride,
                                   int width, int num_comp, uint16_t bg, byte *gs_restrict linebuf);
 void gx_blend_image_buffer(byte *buf_ptr, int width, int height,
-    int rowstride, int planestride, int num_comp, byte bg);
+                      int rowstride, int planestride, int num_comp, byte bg);
 void gx_blend_image_buffer16(byte *buf_ptr, int width, int height,
     int rowstride, int planestride, int num_comp, uint16_t bg);
-void gx_blend_image_buffer8to16(const byte *buf_ptr, unsigned short *buf_ptr_out, int width, int height,
-        int rowstride, int planestride, int num_comp, byte bg);
+void gx_blend_image_buffer8to16(const byte *buf_ptr, unsigned short *buf_ptr_out,
+    int width, int height, int rowstride, int planestride, int num_comp, byte bg);
 int gx_put_blended_image_cmykspot(gx_device *target, byte *buf_ptr,
                       int planestride, int rowstride,
                       int x0, int y0, int width, int height, int num_comp, uint16_t bg,
@@ -413,7 +413,7 @@ void dump_raw_buffer(const gs_memory_t *mem,
                      int num_rows, int width, int n_chan,
                      int plane_stride, int rowstride,
                      char filename[],const byte *Buffer, bool deep);
-void dump_raw_buffer_be(const gs_memory_t *mem, 
+void dump_raw_buffer_be(const gs_memory_t *mem,
                         int num_rows, int width, int n_chan,
                         int plane_stride, int rowstride,
                         char filename[],const byte *Buffer, bool deep);
